+++ src/lib.rs
@@ -334,7 +334,7 @@
 //! [`examples/custom_filter.rs`]: https://github.com/yaahc/color-eyre/blob/master/examples/custom_filter.rs
 //! [`examples/custom_section.rs`]: https://github.com/yaahc/color-eyre/blob/master/examples/custom_section.rs
 //! [`examples/multiple_errors.rs`]: https://github.com/yaahc/color-eyre/blob/master/examples/multiple_errors.rs
-#![doc(html_root_url = "https://docs.rs/color-eyre/0.5.4")]
+#![doc(html_root_url = "https://docs.rs/color-eyre/0.5.5")]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![warn(
     missing_docs,
@@ -405,6 +405,18 @@ pub struct Handler {
     #[cfg(feature = "issue-url")]
     issue_metadata:
         std::sync::Arc<Vec<(String, Box<dyn std::fmt::Display + Send + Sync + 'static>)>>,
+    #[cfg(feature = "issue-url")]
+    issue_filter: std::sync::Arc<config::IssueFilterCallback>,
+}
+
+/// The kind of type erased error being reported
+#[cfg(feature = "issue-url")]
+#[cfg_attr(docsrs, doc(cfg(feature = "issue-url")))]
+pub enum ErrorKind<'a> {
+    /// A non recoverable error aka `panic!`
+    NonRecoverable(&'a dyn std::any::Any),
+    /// A recoverable error aka `impl std::error::Error`
+    Recoverable(&'a (dyn std::error::Error + 'static)),
 }
 
 static CONFIG: OnceCell<config::PanicHook> = OnceCell::new();